edit client -> device2
