+++ device2/src/cfile/speaker.c
